Renamed discussion staff menu to staff courses menu
